Moves SnowLayer source and header files to /src and /include
