+++ src/snowsoil/TemperatureUpdator.h
@@ -12,7 +12,7 @@
 
 #include "../../include/Ground.h"
 #include "../../include/Layer.h"
-#include "../ecodomain/layer/SnowLayer.h"
+#include "../../include/SnowLayer.h"
 #include "../ecodomain/layer/SoilLayer.h"
 
 #include "../inc/physicalconst.h"
